grid stuff updated

+++ Assets/Script/Tile.cs
@@ -3,6 +3,6 @@ using System;
 public class Tile
 {
     public bool isAlive;
-    public UInt16 X;
-    public UInt16 Y;
+    public int X;
+    public int Y;
 }
